Move default and known app metas to common place
